esquema MySQL e integridad relacional

+++ backend/src/main/java/com/medikids/medikids/expose/model/request/PagoRequest.java
@@ -4,13 +4,13 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
-import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Setter
 @Getter
 public class PagoRequest implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private double monto;
+    private BigDecimal monto;
     private String metodo_pago;
 }
